fixed AnimeFrenzy
The website has changed a lot since the creation of the provider, I use the old website for the scraping so I don't know how long this fix will last, but since its the old design it means it wont get any changes. Only change I can think of is the deletion of the old webpage portal.